backend/zewadi/requirements.txt
Django>=6.0,<7.0
djangorestframework>=3.15
djangorestframework-simplejwt>=5.4
django-cors-headers>=4.4
python-dotenv>=1.0
Pillow>=10.0
psycopg2-binary>=2.9
gunicorn>=22.0
